+++ include/GB/memory/OAM.h
@@ -14,8 +14,8 @@ public:
     OAM();
     ~OAM();
 
-    u8 readMemoryU8(u16 address);
-    void writeMemoryU8(u16 address, u8 data);
+    u8 readMemoryU8(u32 address);
+    void writeMemoryU8(u32 address, u8 data);
 };
 
 // https://gbdev.io/pandocs/Memory_Map.html#fea0feff-range
