+++ include/GB/memory/ROM.h
@@ -22,8 +22,8 @@ public:
 
     void writeROM(u8* rom, u32 size);
 
-    u8 readMemoryU8(u16 address);
-    void writeMemoryU8(u16 address, u8 data);
+    u8 readMemoryU8(u32 address);
+    void writeMemoryU8(u32 address, u8 data);
 };
 
 #endif // ROM_H
