+++ include/GB/memory/VRAM.h
@@ -16,8 +16,8 @@ public:
     VRAM();
     ~VRAM();
 
-    u8 readMemoryU8(u16 address);
-    void writeMemoryU8(u16 address, u8 data);
+    u8 readMemoryU8(u32 address);
+    void writeMemoryU8(u32 address, u8 data);
 };
 
 #endif // VRAM_H
