+++ include/GB/memory/cartridge.h
@@ -27,10 +27,10 @@ public:
 
     void printCartridgeData();
 
-    u8 readMemoryU8(u16 address);
-    void writeMemoryU8(u16 address, u8 data);
+    u8 readMemoryU8(u32 address);
+    void writeMemoryU8(u32 address, u8 data);
     
-    bool isAddressInRange(u16 address);
+    bool isAddressInRange(u32 address);
 };
 
 #endif // CARTRIDGE_H
