+++ src/GB/memory/VRAM.cpp
@@ -7,7 +7,7 @@ VRAM::~VRAM() {}
 
 // *****
 
-u8 VRAM::readMemoryU8(u16 address) {
+u8 VRAM::readMemoryU8(u32 address) {
     if (!this->isAddressInRange(address)) {
         std::cerr << "Out of bounds VRAM READ:" << std::endl;
         std::cerr << "Address:  0x" << std::hex << address << std::endl;
@@ -18,7 +18,7 @@ u8 VRAM::readMemoryU8(u16 address) {
     return this->data[address - VRAM_RANGE_FROM];
 }
 
-void VRAM::writeMemoryU8(u16 address, u8 value) {
+void VRAM::writeMemoryU8(u32 address, u8 value) {
     if (!this->isAddressInRange(address)) {
         std::cerr << "Out of bounds VRAM WRITE:" << std::endl;
         std::cerr << "Address:  0x" << std::hex << address << std::endl;
